requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cachetools==5.2.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
databricks-cli==0.17.4
debugpy==1.6.5
decorator==4.4.2
dnspython==2.3.0
docker==6.0.1
docker-pycreds==0.4.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.2
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
gcsfs==2022.11.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
gunicorn==20.1.0
huggingface-hub==0.12.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
importlib-metadata==5.2.0
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyterlab-widgets==3.0.5
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mlflow==1.30.0
more-itertools==9.0.0
moviepy==1.0.3
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20230124
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
wcwidth==0.2.6
websocket-client==1.5.0
Werkzeug==2.2.2
widgetsnbextension==4.0.5
yarl==1.8.2
zipp==3.11.0
